Dividing code
