+++ src/awesome_package/models.py
@@ -1,0 +1,103 @@
+import os
+import tarfile
+
+import numpy as np
+import pandas as pd
+from scipy.stats import randint
+from six.moves import urllib
+from sklearn.ensemble import RandomForestRegressor
+from sklearn.impute import SimpleImputer
+from sklearn.linear_model import LinearRegression
+from sklearn.metrics import mean_absolute_error, mean_squared_error
+from sklearn.model_selection import (
+    GridSearchCV,
+    RandomizedSearchCV,
+    StratifiedShuffleSplit,
+    train_test_split,
+)
+from sklearn.tree import DecisionTreeRegressor
+def linear():
+    lin_reg = LinearRegression()
+    lin_reg.fit(housing_prepared, housing_labels)
+
+    housing_predictions = lin_reg.predict(housing_prepared)
+    lin_mse = mean_squared_error(housing_labels, housing_predictions)
+    lin_rmse = np.sqrt(lin_mse)
+    lin_rmse
+
+    lin_mae = mean_absolute_error(housing_labels, housing_predictions)
+    lin_mae
+
+    tree_reg = DecisionTreeRegressor(random_state=42)
+    tree_reg.fit(housing_prepared, housing_labels)
+
+    housing_predictions = tree_reg.predict(housing_prepared)
+    tree_mse = mean_squared_error(housing_labels, housing_predictions)
+    tree_rmse = np.sqrt(tree_mse)
+    tree_rmse
+
+    param_distribs = {
+        "n_estimators": randint(low=1, high=200),
+        "max_features": randint(low=1, high=8),
+    }
+
+    forest_reg = RandomForestRegressor(random_state=42)
+    rnd_search = RandomizedSearchCV(
+        forest_reg,
+        param_distributions=param_distribs,
+        n_iter=10,
+        cv=5,
+        scoring="neg_mean_squared_error",
+        random_state=42,
+    )
+    rnd_search.fit(housing_prepared, housing_labels)
+    cvres = rnd_search.cv_results_
+    for mean_score, params in zip(cvres["mean_test_score"], cvres["params"]):
+        print(np.sqrt(-mean_score), params)
+
+    param_grid = [
+    # try 12 (3×4) combinations of hyperparameters
+        {"n_estimators": [3, 10, 30], "max_features": [2, 4, 6, 8]},
+    # then try 6 (2×3) combinations with bootstrap set as False
+        {"bootstrap": [False], "n_estimators": [3, 10], "max_features": [2, 3, 4]},
+    ]
+
+    forest_reg = RandomForestRegressor(random_state=42)
+# train across 5 folds, that's a total of (12+6)*5=90 rounds of training
+    grid_search = GridSearchCV(
+        forest_reg,
+        param_grid,
+        cv=5,
+        scoring="neg_mean_squared_error",
+        return_train_score=True,
+    )
+    grid_search.fit(housing_prepared, housing_labels)
+
+    grid_search.best_params_
+    cvres = grid_search.cv_results_
+    for mean_score, params in zip(cvres["mean_test_score"], cvres["params"]):
+        print(np.sqrt(-mean_score), params)
+
+    feature_importances = grid_search.best_estimator_.feature_importances_
+    sorted(zip(feature_importances, housing_prepared.columns), reverse=True)
+
+    final_model = grid_search.best_estimator_
+
+    X_test = strat_test_set.drop("median_house_value", axis=1)
+    y_test = strat_test_set["median_house_value"].copy()
+
+    X_test_num = X_test.drop("ocean_proximity", axis=1)
+    X_test_prepared = imputer.transform(X_test_num)
+    X_test_prepared = pd.DataFrame(
+        X_test_prepared, columns=X_test_num.columns, index=X_test.index
+    )
+    X_test_prepared["rooms_per_household"] = (
+        X_test_prepared["total_rooms"] / X_test_prepared["households"]
+    )
+    X_test_prepared["bedrooms_per_room"] = (
+        X_test_prepared["total_bedrooms"] / X_test_prepared["total_rooms"]
+    )
+    X_test_prepared["population_per_household"] = (
+        X_test_prepared["population"] / X_test_prepared["households"]
+    )
+    return X_test_prepared,y_test
